experiment-electrum/requirements.txt
Electrum==2.7.12
again==1.2.21
argparse==1.2.1
cycler==0.10.0
dnspython==1.15.0
ecdsa==0.13
jsonrpclib==0.1.7
jsonstreamer==1.3.6
matplotlib==1.5.3
numpy==1.11.3
pbkdf2==1.3
progressbar2==3.12.0
protobuf==3.1.0.post1
pycrypto==2.6.1
pyparsing==2.1.10
python-dateutil==2.6.0
python-utils==2.0.1
pytz==2016.10
qrcode==5.3
requests==2.12.4
six==1.10.0
slowaes==0.1a1
wsgiref==0.1.2
